clean this up and move things to better locations
