+++ policy/uploader/__init__.py
@@ -1,5 +1,14 @@
 #!/usr/bin/python
 """
-CherryPy Uploader object classes
+CherryPy Uploader Policy object class
 """
-from policy import UploaderPolicy
+
+# pylint: disable=too-few-public-methods
+class UploaderPolicy(object):
+    """
+    CherryPy root object class
+
+    not exposed by default the base objects are exposed
+    """
+    exposed = True
+# pylint: enable=too-few-public-methods
