--- policy/uploader/policy.py
@@ -1,15 +1,0 @@
-#!/usr/bin/python
-"""
-CherryPy Uploader Policy object class
-"""
-
-# pylint: disable=too-few-public-methods
-class UploaderPolicy(object):
-    """
-    CherryPy root object class
-
-    not exposed by default the base objects are exposed
-    """
-    exposed = True
-# pylint: enable=too-few-public-methods
-
